Move things around more
