+++ src/mx_bluesky/common/external_interaction/callbacks/common/aperture_change_callback.py
