+++ src/mx_bluesky/common/external_interaction/callbacks/common/grid_detection_callback.py
