+++ tests/unit_tests/common/external_interaction/callbacks/test_plan_reactive_callback.py
@@ -6,7 +6,7 @@ import pytest
 from bluesky.run_engine import RunEngine
 from event_model.documents import Event, EventDescriptor, RunStart, RunStop
 
-from mx_bluesky.common.external_interaction.callbacks.plan_reactive_callback import (
+from mx_bluesky.common.external_interaction.callbacks.common.plan_reactive_callback import (
     PlanReactiveCallback,
 )
