+++ tests/unit_tests/common/external_interaction/callbacks/test_zocalo_handler.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock, call, patch
 import pytest
 from dodal.devices.zocalo import ZocaloStartInfo
 
-from mx_bluesky.common.external_interaction.callbacks.zocalo_callback import (
+from mx_bluesky.common.external_interaction.callbacks.common.zocalo_callback import (
     ZocaloCallback,
 )
 from mx_bluesky.common.external_interaction.ispyb.ispyb_store import (
@@ -66,7 +66,7 @@ class TestZocaloHandler:
             )
 
     @patch(
-        "mx_bluesky.common.external_interaction.callbacks.zocalo_callback.ZocaloTrigger",
+        "mx_bluesky.common.external_interaction.callbacks.common.zocalo_callback.ZocaloTrigger",
         autospec=True,
     )
     def test_handler_inits_zocalo_trigger_on_right_plan(self, zocalo_trigger):
@@ -82,7 +82,7 @@ class TestZocaloHandler:
         assert zocalo_handler.zocalo_interactor is not None
 
     @patch(
-        "mx_bluesky.common.external_interaction.callbacks.zocalo_callback.ZocaloTrigger",
+        "mx_bluesky.common.external_interaction.callbacks.common.zocalo_callback.ZocaloTrigger",
         autospec=True,
     )
     @patch(
